Refactor schema into domain modules
