+++ lib/db/schema/index.ts
@@ -1,0 +1,8 @@
+export * from "./auth";
+export * from "./chat";
+export * from "./projects";
+export * from "./source-material";
+export * from "./entities";
+export * from "./outlines";
+export * from "./generation";
+export * as legacySchema from "./legacy";
